modified galago types to use longs instead of integers for document ids

+++ contrib/src/main/java/org/lemurproject/galago/contrib/index/InvertedSketchIndexWriter.java
@@ -97,7 +97,7 @@ public class InvertedSketchIndexWriter implements
   }
 
   @Override
-  public void processDocument(int document) throws IOException {
+  public void processDocument(long document) throws IOException {
     invertedList.addDocument(document);
   }
